+++ spec/spec_helper.rb
@@ -8,6 +8,7 @@ require 'logger'
 require 'rails/all'
 require 'js-routes'
 require 'active_support/core_ext/hash/slice'
+require 'active_support/core_ext/object/to_query'
 
 unless ENV['CI']
   code = system("yarn build")
@@ -57,6 +58,7 @@ rescue MiniRacer::RuntimeError => e
 end
 
 def evallib(**options)
+  options[:include_undefined_query_parameters] ||= false
   evaljs(JsRoutes.generate(**options), filename: 'lib/routes.js')
 end
 
@@ -126,6 +128,8 @@ RSpec.configure do |config|
   end
 
   config.before :each do
+    JsRoutes.configuration.include_undefined_query_parameters = true
+
     log = proc do |*values|
       puts values.map(&:inspect).join(", ")
     end